# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\"Starter code for Go application with Gin, System Tray, Gorm, Air, Swagger\"",
        "title": "gin-systray-starter",
        "contact": {
            "name": "Github Issues",
            "url": "https://github.com/jinyaoMa/gin-systray-starter/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/jinyaoMa/gin-systray-starter/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/test/": {
            "get": {
                "description": "Test Controller",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Before Authorization"
                ],
                "summary": "TestController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Msg",
                        "name": "msg",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ ok , data }"
                    },
                    "400": {
                        "description": "binding error"
                    },
                    "404": {
                        "description": "{ error }"
                    }
                }
            }
        },
        "/test/checkToken/": {
            "get": {
                "security": [
                    {
                        "BearerIdAuth": []
                    }
                ],
                "description": "Test Jwt, check token",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "After Authorization"
                ],
                "summary": "TestJwtCheckToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ ok , claims }"
                    },
                    "401": {
                        "description": "Auth failed"
                    },
                    "404": {
                        "description": "{ error }"
                    }
                }
            }
        },
        "/test/getToken/": {
            "get": {
                "description": "Test Jwt, get token",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Before Authorization"
                ],
                "summary": "TestJwtGetToken",
                "responses": {
                    "200": {
                        "description": "{ ok , token }"
                    },
                    "404": {
                        "description": "{ error }"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerIdAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}